requirements.txt
torch
ultralytics
fastapi
uvicorn
loguru
requests
pillow
opencv-python
flake8
pylint
streamlit
pandas
matplotlib
seaborn
python-multipart
python-dotenv